Created View Model Changed Package names
