--- app/src/main/java/com/example/theyelpapp/data/model/restaurant/Category.kt
@@ -1,15 +1,0 @@
-package com.example.theyelpapp.data.model.restaurant
-
-
-import com.google.gson.annotations.SerializedName
-
-/**
- * Data class that defines Category
- */
-
-data class Category(
-    @SerializedName("alias")
-    val alias: String? = null,
-    @SerializedName("title")
-    val title: String? = null
-)
