--- app/src/main/java/com/example/theyelpapp/data/model/restaurant/Coordinates.kt
@@ -1,15 +1,0 @@
-package com.example.theyelpapp.data.model.restaurant
-
-
-import com.google.gson.annotations.SerializedName
-
-/**
- * Data class that defines Coordinates
- */
-
-data class Coordinates(
-    @SerializedName("latitude")
-    val latitude: Double? = null,
-    @SerializedName("longitude")
-    val longitude: Double? = null
-)
