--- app/src/main/java/com/example/theyelpapp/data/model/restaurant/RestaurantsResponse.kt
@@ -1,17 +1,0 @@
-package com.example.theyelpapp.data.model.restaurant
-
-
-import com.google.gson.annotations.SerializedName
-
-/**
- * Data class that defines Restaurants Response
- */
-
-data class RestaurantsResponse(
-    @SerializedName("businesses")
-    val businesses: List<Businesses>? = null,
-//    @SerializedName("region")
-//    val region: Region? = null,
-    @SerializedName("total")
-    val total: Int? = null
-)
