+++ app/src/main/java/com/example/theyelpapp/datalayer/model/restaurant/Region.kt
@@ -1,0 +1,13 @@
+package com.example.theyelpapp.datalayer.model.restaurant
+
+
+import com.google.gson.annotations.SerializedName
+
+/**
+ * Data class that defines Region
+ */
+
+data class Region(
+    @SerializedName("center")
+    val center: Center? = null
+)
