+++ app/src/main/java/com/example/theyelpapp/di/ApplicationModule.kt
@@ -2,6 +2,12 @@ package com.example.theyelpapp.di
 
 import android.content.Context
 import android.net.ConnectivityManager
+import com.example.theyelpapp.datalayer.service.NetworkRepository
+import com.example.theyelpapp.usecaseslayer.GetLocationUseCase
+import com.example.theyelpapp.usecaseslayer.GetRatingsUseCase
+import com.example.theyelpapp.usecaseslayer.GetRestaurantsUseCase
+import com.example.theyelpapp.usecaseslayer.YelpUseCases
+import com.example.theyelpapp.utils.NetworkState
 import com.google.android.gms.location.FusedLocationProviderClient
 import com.google.android.gms.location.LocationServices
 import dagger.Module
@@ -9,21 +15,37 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
 class ApplicationModule {
 
     @Provides
+    @Singleton
     fun providesConnectivityManager(
         @ApplicationContext context: Context
     ): ConnectivityManager =
         context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
 
     @Provides
+    @Singleton
     fun providesFusedLocation(
         @ApplicationContext context: Context
     ): FusedLocationProviderClient =
         LocationServices.getFusedLocationProviderClient(context)
+
+    @Provides
+    @Singleton
+    fun providesYelpUseCases(
+        fusedLocation: FusedLocationProviderClient,
+        networkRepository: NetworkRepository,
+        networkState: NetworkState
+    ): YelpUseCases =
+        YelpUseCases(
+            GetLocationUseCase(fusedLocation),
+            GetRatingsUseCase(networkRepository,networkState),
+            GetRestaurantsUseCase(networkRepository,networkState)
+        )
 
 }
