+++ app/src/main/java/com/example/theyelpapp/usecaseslayer/YelpUseCases.kt
@@ -1,0 +1,7 @@
+package com.example.theyelpapp.usecaseslayer
+
+data class YelpUseCases(
+    val getLocation: GetLocationUseCase,
+    val getRatings: GetRatingsUseCase,
+    val getRestaurants: GetRestaurantsUseCase
+)
